(MVP DAY2) : refactored App container
