+++ friends/src/containers/App/index.jsx
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { Route, Switch, withRouter } from 'react-router-dom';
-import Header from './components/Header';
-import Form from './components/Form';
+import Header from '../../components/Header';
+import Form from '../../components/Form';
 import './App.css';
-import FriendList from './components/FriendList';
+import FriendList from '../../components/FriendList';
 
 class App extends Component {
